backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pywinpty==2.0.14
websockets==14.1
pydantic-settings==2.7.1
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
